requirements.txt
# Core Dependencies
streamlit>=1.54.0
langchain>=1.2.10
langchain-openai>=1.1.10
langchain-google-genai>=4.2.0
langchain-community>=0.4.1
langchain-core>=1.2.13
langchain-text-splitters>=1.1.0

# Vector Database
faiss-cpu>=1.13.2

# PDF Processing
PyPDF2>=3.0.1

# Web Scraping
requests>=2.32.5
beautifulsoup4>=4.14.3

# Environment Management
python-dotenv>=1.2.1

# OpenAI
openai>=2.21.0

# Utilities
tiktoken>=0.12.0